+++ include/chain_planner/target_modes/joint_modes/JointStepWaveMode.h
@@ -19,7 +19,7 @@ private:
 
 
 public:
-    JointStepWaveMode(const int _amplitude, const double _period, const double _duration, boost::function<void(double)> _setAngleFunction)
+    JointStepWaveMode(const double _amplitude, const double _period, const double _duration, boost::function<void(double)> _setAngleFunction)
     : DiscreteMode(0.5*_period, _duration, boost::bind(&JointStepWaveMode::callback, this))
     , angle(_amplitude)
     , setAngleFunction(_setAngleFunction)
